requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.2.0
cached-property==1.5.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.3
Django==3.2.6
django-cors-headers==3.8.0
django-debug-toolbar==3.2.2
django-elasticsearch-dsl==7.2.1
django-filter==2.4.0
django-ipware==4.0.0
django-js-asset==1.2.2
django-mptt==0.13.2
django-tags-input==5.1.0
djangorestframework==3.12.4
djangorestframework-recursive==0.1.2
easy-thumbnails==2.7.1
elasticsearch==7.15.1
elasticsearch-dsl==7.4.0
expiringdict==1.2.1
frozenlist==1.2.0
geoip2==4.5.0
gTTS==2.2.3
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.7.1
iniconfig==1.1.1
mailru-im-bot==0.0.20
Markdown==3.3.4
maxminddb==2.2.0
monotonic==1.6
multidict==5.2.0
packaging==21.0
Pillow==8.3.1
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
Pygments==2.11.2
pyparsing==2.4.7
pytest==6.2.5
python-baseconv==1.2.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.7
yarl==1.7.2
zipp==3.5.0
